Create down_vip_upon_min_active_servers.py
This script (can be used as an external health monitor) will read the min_active_members option of a defined Service Group then it will check the UP status of all the servers in that Service Group (specifically the server ports in that service group).  It will return FAIL (1) if the UP server count is less than min_active_members.  It will return SUCCESS (0) if min_active_member is disabled or there are more UP servers than min_active_member value.